removed: calendar v1 in favor of new calendar implementation.
