+++ packages/luna/src/components/Calendar/Events/index.ts
@@ -1,0 +1,2 @@
+export { LessonSlot } from "@/components/Calendar/Events/LessonSlot/LessonSlot";
+export { AvailabilitySlot } from "@/components/Calendar/Events/AvailabilitySlots/AvailabilitySlots";
