+++ packages/luna/src/components/Calendar/Events/shared/MemberAvatar.tsx
