+++ packages/luna/src/components/Calendar/index.ts
@@ -1,2 +1,3 @@
 export { Calendar } from "@/components/Calendar/Calendar";
 export type { IEvent as Event } from "@/components/Calendar/types";
+export { asHourId, isHourId } from "@/components/Calendar/Hours";
