--- packages/luna/src/components/Calendar/v2/Events/index.ts
@@ -1,2 +1,0 @@
-export { LessonSlot } from "@/components/Calendar/v2/Events/LessonSlot/LessonSlot";
-export { AvailabilitySlot } from "@/components/Calendar/v2/Events/AvailabilitySlots/AvailabilitySlots";
