--- packages/luna/src/components/Calendar/v2/Events/shared/index.tsx
@@ -1,3 +1,0 @@
-export { EventSpan } from "@/components/Calendar/v2/Events/shared/EventSpan";
-export { MemberAvatar } from "@/components/Calendar/v2/Events/shared/MemberAvatar";
-export { Card } from "@/components/Calendar/v2/Events/shared/Card";
