--- packages/luna/src/components/Calendar/v2/index.ts
@@ -1,2 +1,0 @@
-export { Calendar } from "@/components/Calendar/v2/Calendar";
-export { asHourId, isHourId } from "@/components/Calendar/v2/Hours";
